+++ src/Content/Src/Sound.cpp
@@ -1,5 +1,5 @@
 ﻿#include "../Include/Sound.h"
-#include "../../System/Include/Logger.h"
+#include "../../Core/Include/Logger.h"
 #include "SFML/Audio/SoundBuffer.hpp"
 #include "SFML/System/Exception.hpp"
