+++ src/Core/Include/Extensions.h
