+++ src/Core/Include/SpawnHelper.h
