+++ src/Core/Src/Quaternion.cpp
