+++ src/Entities/Src/Bullets.cpp
@@ -2,11 +2,11 @@
 #include "../../GameRoot.h"
 #include "../../Content/Include/GaussianBlur.h"
 #include "../../Content/Include/Sound.h"
-#include "../../Grid/Grid.h"
-#include "../../Particles/Particles.h"
-#include "../../System/Include/Extensions.h"
-#include "../../System/Include/Quaternion.h"
-#include "../../System/Include/RandomVector.h"
+#include "../../Systems/Include/Grid.h"
+#include "../../Systems/Include/Particles.h"
+#include "../../Core/Include/Extensions.h"
+#include "../../Core/Include/Quaternion.h"
+#include "../../Core/Include/RandomVector.h"
 
 
 Bullets::Bullets()
