+++ src/Entities/Src/Enemies.cpp
@@ -3,13 +3,13 @@
 #include "../../GameRoot.h"
 #include "../../Content/Include/GaussianBlur.h"
 #include "../../Content/Include/Sound.h"
-#include "../../Particles/Particles.h"
+#include "../../Systems/Include/Particles.h"
 #include "../Include/Player/Buffs.h"
 #include "../Include/Player/PlayerStatus.h"
-#include "../../System/Include/ColorPicker.h"
-#include "../../System/Include/Extensions.h"
-#include "../../System/Include/RandomVector.h"
-#include "../../System/Include/SpawnHelper.h"
+#include "../../Core/Include/ColorPicker.h"
+#include "../../Core/Include/Extensions.h"
+#include "../../Core/Include/RandomVector.h"
+#include "../../Core/Include/SpawnHelper.h"
 #include "../../UserInterface/Include/FloatingKillTexts.h"
 #include "../Include/Player/PlayerShip.h"
 #include "SFML/Graphics/Text.hpp"
