+++ src/Entities/Src/Player/PlayerShip.cpp
@@ -5,10 +5,10 @@
 #include "../../../Content/Include/Art.h"
 #include "../../../Content/Include/GaussianBlur.h"
 #include "../../../Input/Include/Input.h"
-#include "../../../Particles/Particles.h"
+#include "../../../Systems/Include/Particles.h"
 #include "../../Include/Player/PlayerStatus.h"
-#include "../../../System/Include/Extensions.h"
-#include "../../../System/Include/RandomVector.h"
+#include "../../../Core/Include/Extensions.h"
+#include "../../../Core/Include/RandomVector.h"
 #include "../../Include/Bullets.h"
 
 PlayerShip::PlayerShip()
