+++ src/GameRoot.cpp
@@ -10,10 +10,10 @@
 #include "Entities/Include/Nukes.h"
 #include "Entities/Include/ShapeKeeper/ShapeKeeper.h"
 #include "Entities/Include/Player/PlayerShip.h"
-#include "Grid/Grid.h"
+#include "Systems/Include/Grid.h"
 #include "Input/Include/Input.h"
-#include "System/Include/Logger.h"
-#include "Particles/Particles.h"
+#include "Core/Include/Logger.h"
+#include "Systems/Include/Particles.h"
 #include "Entities/Include/Player/Buffs.h"
 #include "Entities/Include/Player/PlayerStatus.h"
 #include "SFML/Graphics/Image.hpp"
