+++ src/Systems/Include/Grid.h
@@ -1,7 +1,7 @@
 ﻿#ifndef GRID_H
 #define GRID_H
 #include <vector>
-#include "../Content/Include/Art.h"
+#include "../../Content/Include/Art.h"
 #include "SFML/Graphics/Color.hpp"
 #include "SFML/System/Vector2.hpp"
 #include "SFML/System/Vector3.hpp"
